refactor: Phase 1 (SQL dialect abstraction) + Phase 3 (Data engine modularization)
Phase 1 — SQL Dialect Abstraction:
- fieldRegistry.ts: single source of truth for event field definitions
- sqlDialect.ts: common interface (buildSelect, buildWhere, getTotalCount)
- postgresDialect.ts / clickHouseDialect.ts: dialect implementations
- queryEvents.ts: unified handler that picks dialect via STORAGE_ADAPTER
- Both PG and CH handlers now thin re-exports from common/queryEvents

Phase 3 — Data Engine Modularization:
- Extract dataQuery module from routes/gRPC/data/query into storage/query/
- Explicit field alias registry with auto-derived cast types from Drizzle metadata
- Thin handler in routes/gRPC/data/query delegates to executeDataQuery/getDataTable
- hasMore support in DataQueryResult
