+++ src/storage/adapter/common/fieldRegistry.ts
@@ -1,0 +1,525 @@
+import type { EventTableName } from "./queryEventsBase";
+
+export interface FieldDef {
+  pgSelect: string | null;
+  pgWhereCol: string | null;
+  pgWhereCast: string;
+  pgAggExpr?: string;
+  chSelect: string | null;
+  chWhere: string | null;
+  chAggExpr?: string;
+  chParamType: string;
+}
+
+export const FIELD_REGISTRY: Record<
+  EventTableName,
+  Record<string, FieldDef>
+> = {
+  basic_usage_events: {
+    eventId: {
+      pgSelect: "event_id::text",
+      pgWhereCol: "event_id",
+      pgWhereCast: "::uuid",
+      chSelect: "event_id",
+      chWhere: "event_id",
+      chParamType: "String",
+    },
+    idempotencyKey: {
+      pgSelect: "idempotency_key",
+      pgWhereCol: "idempotency_key",
+      pgWhereCast: "",
+      chSelect: "idempotency_key",
+      chWhere: "idempotency_key",
+      chParamType: "String",
+    },
+    mode: {
+      pgSelect: "mode",
+      pgWhereCol: "mode",
+      pgWhereCast: "",
+      chSelect: "mode",
+      chWhere: "mode",
+      chParamType: "String",
+    },
+    eventType: {
+      pgSelect: "'BASIC_USAGE'",
+      pgWhereCol: null,
+      pgWhereCast: "",
+      chSelect: "'BASIC_USAGE'",
+      chWhere: null,
+      chParamType: "String",
+    },
+    userId: {
+      pgSelect: "user_id::text",
+      pgWhereCol: "user_id",
+      pgWhereCast: "",
+      chSelect: "user_id",
+      chWhere: "user_id",
+      chParamType: "String",
+    },
+    apiKeyId: {
+      pgSelect: "api_key_id::text",
+      pgWhereCol: "api_key_id",
+      pgWhereCast: "",
+      chSelect: "api_key_id",
+      chWhere: "api_key_id",
+      chParamType: "String",
+    },
+    reportedTimestamp: {
+      pgSelect: "reported_timestamp::text",
+      pgWhereCol: "reported_timestamp",
+      pgWhereCast: "::timestamptz",
+      chSelect: "toString(reported_timestamp)",
+      chWhere: "reported_timestamp",
+      chParamType: "DateTime64(3, 'UTC')",
+    },
+    ingestedTimestamp: {
+      pgSelect: "ingested_timestamp::text",
+      pgWhereCol: "ingested_timestamp",
+      pgWhereCast: "::timestamptz",
+      chSelect: "toString(ingested_timestamp)",
+      chWhere: "ingested_timestamp",
+      chParamType: "DateTime64(3, 'UTC')",
+    },
+    basicUsageType: {
+      pgSelect: "type",
+      pgWhereCol: "type",
+      pgWhereCast: "",
+      chSelect: "type",
+      chWhere: "type",
+      chParamType: "String",
+    },
+    debitAmount: {
+      pgSelect: "debit_amount::text",
+      pgWhereCol: "debit_amount",
+      pgWhereCast: "::bigint",
+      chSelect: "toString(debit_amount)",
+      chWhere: "debit_amount",
+      chParamType: "Int64",
+    },
+    model: {
+      pgSelect: null,
+      pgWhereCol: null,
+      pgWhereCast: "",
+      chSelect: null,
+      chWhere: null,
+      chParamType: "String",
+    },
+    inputTokens: {
+      pgSelect: null,
+      pgWhereCol: null,
+      pgWhereCast: "",
+      chSelect: null,
+      chWhere: null,
+      chParamType: "Int64",
+    },
+    outputTokens: {
+      pgSelect: null,
+      pgWhereCol: null,
+      pgWhereCast: "",
+      chSelect: null,
+      chWhere: null,
+      chParamType: "Int64",
+    },
+    inputDebitAmount: {
+      pgSelect: null,
+      pgWhereCol: null,
+      pgWhereCast: "",
+      chSelect: null,
+      chWhere: null,
+      chParamType: "Int64",
+    },
+    outputDebitAmount: {
+      pgSelect: null,
+      pgWhereCol: null,
+      pgWhereCast: "",
+      chSelect: null,
+      chWhere: null,
+      chParamType: "Int64",
+    },
+    inputCacheTokens: {
+      pgSelect: null,
+      pgWhereCol: null,
+      pgWhereCast: "",
+      chSelect: null,
+      chWhere: null,
+      chParamType: "Int64",
+    },
+    inputCacheDebitAmount: {
+      pgSelect: null,
+      pgWhereCol: null,
+      pgWhereCast: "",
+      chSelect: null,
+      chWhere: null,
+      chParamType: "Int64",
+    },
+    creditAmount: {
+      pgSelect: null,
+      pgWhereCol: null,
+      pgWhereCast: "",
+      chSelect: null,
+      chWhere: null,
+      chParamType: "Int64",
+    },
+    provider: {
+      pgSelect: null,
+      pgWhereCol: null,
+      pgWhereCast: "",
+      chSelect: null,
+      chWhere: null,
+      chParamType: "String",
+    },
+    metadata: {
+      pgSelect: "metadata::text",
+      pgWhereCol: null,
+      pgWhereCast: "",
+      chSelect: null,
+      chWhere: null,
+      chParamType: "String",
+    },
+  },
+  ai_token_usage_events: {
+    eventId: {
+      pgSelect: "event_id::text",
+      pgWhereCol: "event_id",
+      pgWhereCast: "::uuid",
+      chSelect: "event_id",
+      chWhere: "event_id",
+      chParamType: "String",
+    },
+    idempotencyKey: {
+      pgSelect: "idempotency_key",
+      pgWhereCol: "idempotency_key",
+      pgWhereCast: "",
+      chSelect: "idempotency_key",
+      chWhere: "idempotency_key",
+      chParamType: "String",
+    },
+    mode: {
+      pgSelect: "mode",
+      pgWhereCol: "mode",
+      pgWhereCast: "",
+      chSelect: "mode",
+      chWhere: "mode",
+      chParamType: "String",
+    },
+    eventType: {
+      pgSelect: "'AI_TOKEN_USAGE'",
+      pgWhereCol: null,
+      pgWhereCast: "",
+      chSelect: "'AI_TOKEN_USAGE'",
+      chWhere: null,
+      chParamType: "String",
+    },
+    userId: {
+      pgSelect: "user_id::text",
+      pgWhereCol: "user_id",
+      pgWhereCast: "",
+      chSelect: "user_id",
+      chWhere: "user_id",
+      chParamType: "String",
+    },
+    apiKeyId: {
+      pgSelect: "api_key_id::text",
+      pgWhereCol: "api_key_id",
+      pgWhereCast: "",
+      chSelect: "api_key_id",
+      chWhere: "api_key_id",
+      chParamType: "String",
+    },
+    reportedTimestamp: {
+      pgSelect: "reported_timestamp::text",
+      pgWhereCol: "reported_timestamp",
+      pgWhereCast: "::timestamptz",
+      chSelect: "toString(reported_timestamp)",
+      chWhere: "reported_timestamp",
+      chParamType: "DateTime64(3, 'UTC')",
+    },
+    ingestedTimestamp: {
+      pgSelect: "ingested_timestamp::text",
+      pgWhereCol: "ingested_timestamp",
+      pgWhereCast: "::timestamptz",
+      chSelect: "toString(ingested_timestamp)",
+      chWhere: "ingested_timestamp",
+      chParamType: "DateTime64(3, 'UTC')",
+    },
+    basicUsageType: {
+      pgSelect: null,
+      pgWhereCol: null,
+      pgWhereCast: "",
+      chSelect: null,
+      chWhere: null,
+      chParamType: "String",
+    },
+    debitAmount: {
+      pgSelect:
+        "(COALESCE((metrics->'debit_amount'->>'input')::integer,0) + COALESCE((metrics->'debit_amount'->>'input_cache')::integer,0) + COALESCE((metrics->'debit_amount'->>'output')::integer,0))::text",
+      pgWhereCol: null,
+      pgWhereCast: "",
+      pgAggExpr:
+        "(COALESCE((metrics->'debit_amount'->>'input')::bigint,0) + COALESCE((metrics->'debit_amount'->>'input_cache')::bigint,0) + COALESCE((metrics->'debit_amount'->>'output')::bigint,0))",
+      chSelect:
+        "toString(JSONExtractInt(metrics, 'debit_amount', 'input') + JSONExtractInt(metrics, 'debit_amount', 'input_cache') + JSONExtractInt(metrics, 'debit_amount', 'output'))",
+      chWhere: null,
+      chAggExpr:
+        "JSONExtractInt(metrics, 'debit_amount', 'input') + JSONExtractInt(metrics, 'debit_amount', 'input_cache') + JSONExtractInt(metrics, 'debit_amount', 'output')",
+      chParamType: "Int64",
+    },
+    model: {
+      pgSelect: "model",
+      pgWhereCol: "model",
+      pgWhereCast: "",
+      chSelect: "model",
+      chWhere: "model",
+      chParamType: "String",
+    },
+    inputTokens: {
+      pgSelect: "(metrics->'tokens'->>'input')::text",
+      pgWhereCol: null,
+      pgWhereCast: "",
+      pgAggExpr: "(metrics->'tokens'->>'input')::bigint",
+      chSelect: "toString(JSONExtractInt(metrics, 'tokens', 'input'))",
+      chWhere: null,
+      chAggExpr: "JSONExtractInt(metrics, 'tokens', 'input')",
+      chParamType: "Int64",
+    },
+    outputTokens: {
+      pgSelect: "(metrics->'tokens'->>'output')::text",
+      pgWhereCol: null,
+      pgWhereCast: "",
+      pgAggExpr: "(metrics->'tokens'->>'output')::bigint",
+      chSelect: "toString(JSONExtractInt(metrics, 'tokens', 'output'))",
+      chWhere: null,
+      chAggExpr: "JSONExtractInt(metrics, 'tokens', 'output')",
+      chParamType: "Int64",
+    },
+    inputDebitAmount: {
+      pgSelect: "(metrics->'debit_amount'->>'input')::text",
+      pgWhereCol: null,
+      pgWhereCast: "",
+      pgAggExpr: "(metrics->'debit_amount'->>'input')::bigint",
+      chSelect: "toString(JSONExtractInt(metrics, 'debit_amount', 'input'))",
+      chWhere: null,
+      chAggExpr: "JSONExtractInt(metrics, 'debit_amount', 'input')",
+      chParamType: "Int64",
+    },
+    outputDebitAmount: {
+      pgSelect: "(metrics->'debit_amount'->>'output')::text",
+      pgWhereCol: null,
+      pgWhereCast: "",
+      pgAggExpr: "(metrics->'debit_amount'->>'output')::bigint",
+      chSelect: "toString(JSONExtractInt(metrics, 'debit_amount', 'output'))",
+      chWhere: null,
+      chAggExpr: "JSONExtractInt(metrics, 'debit_amount', 'output')",
+      chParamType: "Int64",
+    },
+    inputCacheTokens: {
+      pgSelect: "(metrics->'tokens'->>'input_cache')::text",
+      pgWhereCol: null,
+      pgWhereCast: "",
+      pgAggExpr: "(metrics->'tokens'->>'input_cache')::bigint",
+      chSelect: "toString(JSONExtractInt(metrics, 'tokens', 'input_cache'))",
+      chWhere: null,
+      chAggExpr: "JSONExtractInt(metrics, 'tokens', 'input_cache')",
+      chParamType: "Int64",
+    },
+    inputCacheDebitAmount: {
+      pgSelect: "(metrics->'debit_amount'->>'input_cache')::text",
+      pgWhereCol: null,
+      pgWhereCast: "",
+      pgAggExpr: "(metrics->'debit_amount'->>'input_cache')::bigint",
+      chSelect:
+        "toString(JSONExtractInt(metrics, 'debit_amount', 'input_cache'))",
+      chWhere: null,
+      chAggExpr: "JSONExtractInt(metrics, 'debit_amount', 'input_cache')",
+      chParamType: "Int64",
+    },
+    creditAmount: {
+      pgSelect: null,
+      pgWhereCol: null,
+      pgWhereCast: "",
+      chSelect: null,
+      chWhere: null,
+      chParamType: "Int64",
+    },
+    provider: {
+      pgSelect: "provider",
+      pgWhereCol: "provider",
+      pgWhereCast: "",
+      chSelect: "provider",
+      chWhere: "provider",
+      chParamType: "String",
+    },
+    metadata: {
+      pgSelect: "metadata::text",
+      pgWhereCol: null,
+      pgWhereCast: "",
+      chSelect: "toString(metadata)",
+      chWhere: null,
+      chParamType: "String",
+    },
+  },
+  payment_events: {
+    eventId: {
+      pgSelect: null,
+      pgWhereCol: null,
+      pgWhereCast: "",
+      chSelect: null,
+      chWhere: null,
+      chParamType: "String",
+    },
+    idempotencyKey: {
+      pgSelect: null,
+      pgWhereCol: null,
+      pgWhereCast: "",
+      chSelect: null,
+      chWhere: null,
+      chParamType: "String",
+    },
+    mode: {
+      pgSelect: "mode",
+      pgWhereCol: "mode",
+      pgWhereCast: "",
+      chSelect: null,
+      chWhere: null,
+      chParamType: "String",
+    },
+    eventType: {
+      pgSelect: "'PAYMENT'",
+      pgWhereCol: null,
+      pgWhereCast: "",
+      chSelect: null,
+      chWhere: null,
+      chParamType: "String",
+    },
+    userId: {
+      pgSelect: "user_id::text",
+      pgWhereCol: "user_id",
+      pgWhereCast: "",
+      chSelect: null,
+      chWhere: null,
+      chParamType: "String",
+    },
+    apiKeyId: {
+      pgSelect: "api_key_id::text",
+      pgWhereCol: "api_key_id",
+      pgWhereCast: "",
+      chSelect: null,
+      chWhere: null,
+      chParamType: "String",
+    },
+    reportedTimestamp: {
+      pgSelect: "reported_timestamp::text",
+      pgWhereCol: "reported_timestamp",
+      pgWhereCast: "::timestamptz",
+      chSelect: null,
+      chWhere: null,
+      chParamType: "DateTime64(3, 'UTC')",
+    },
+    ingestedTimestamp: {
+      pgSelect: "ingested_timestamp::text",
+      pgWhereCol: "ingested_timestamp",
+      pgWhereCast: "::timestamptz",
+      chSelect: null,
+      chWhere: null,
+      chParamType: "DateTime64(3, 'UTC')",
+    },
+    basicUsageType: {
+      pgSelect: null,
+      pgWhereCol: null,
+      pgWhereCast: "",
+      chSelect: null,
+      chWhere: null,
+      chParamType: "String",
+    },
+    debitAmount: {
+      pgSelect: null,
+      pgWhereCol: null,
+      pgWhereCast: "",
+      chSelect: null,
+      chWhere: null,
+      chParamType: "Int64",
+    },
+    model: {
+      pgSelect: null,
+      pgWhereCol: null,
+      pgWhereCast: "",
+      chSelect: null,
+      chWhere: null,
+      chParamType: "String",
+    },
+    inputTokens: {
+      pgSelect: null,
+      pgWhereCol: null,
+      pgWhereCast: "",
+      chSelect: null,
+      chWhere: null,
+      chParamType: "Int64",
+    },
+    outputTokens: {
+      pgSelect: null,
+      pgWhereCol: null,
+      pgWhereCast: "",
+      chSelect: null,
+      chWhere: null,
+      chParamType: "Int64",
+    },
+    inputDebitAmount: {
+      pgSelect: null,
+      pgWhereCol: null,
+      pgWhereCast: "",
+      chSelect: null,
+      chWhere: null,
+      chParamType: "Int64",
+    },
+    outputDebitAmount: {
+      pgSelect: null,
+      pgWhereCol: null,
+      pgWhereCast: "",
+      chSelect: null,
+      chWhere: null,
+      chParamType: "Int64",
+    },
+    inputCacheTokens: {
+      pgSelect: null,
+      pgWhereCol: null,
+      pgWhereCast: "",
+      chSelect: null,
+      chWhere: null,
+      chParamType: "Int64",
+    },
+    inputCacheDebitAmount: {
+      pgSelect: null,
+      pgWhereCol: null,
+      pgWhereCast: "",
+      chSelect: null,
+      chWhere: null,
+      chParamType: "Int64",
+    },
+    creditAmount: {
+      pgSelect: "credit_amount::text",
+      pgWhereCol: "credit_amount",
+      pgWhereCast: "::bigint",
+      chSelect: null,
+      chWhere: null,
+      chParamType: "Int64",
+    },
+    provider: {
+      pgSelect: null,
+      pgWhereCol: null,
+      pgWhereCast: "",
+      chSelect: null,
+      chWhere: null,
+      chParamType: "String",
+    },
+    metadata: {
+      pgSelect: null,
+      pgWhereCol: null,
+      pgWhereCast: "",
+      chSelect: null,
+      chWhere: null,
+      chParamType: "String",
+    },
+  },
+};
+
+export const OUTPUT_FIELDS = Object.keys(FIELD_REGISTRY.basic_usage_events!);
